feat: add writing secrets feature in terraform provider
